insertar pedido listo

+++ ServidorJAVA-ClienteCSharp/ServidorJAVA/CrewmanSystemController/src/pe/edu/pucp/CrewmanSystem/dao/LineaPedidoDAO.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 
 public interface LineaPedidoDAO
 {
-    double insertar(LineaPedido lineaPedido);
+    int insertar(LineaPedido lineaPedido);
     int actualizar(LineaPedido lineaPedido);
     int eliminar(int idLineaPedido);
     ArrayList<LineaPedido> listar(int idPedido);
